requirements.txt
python-dotenv>=1.0.1
openai==1.63.2
aiohttp>=3.9.0
aiofiles>=23.2.1
tiktoken>=0.5.0
firecrawl-py>=1.11.1
typer>=0.9.0
prompt-toolkit>=3.0.0
pydantic==2.10.6